box: improve fselect
fselect is designed to make selects results in console better
readable for human eyes. The usage is expected to be rare and
by developers only so there's no performance requirements.

Now the API of fselect is:
s:fselect(<key>, <select+fselect option map>, <fselect option map>)

But sometimes there are cases when only some columns are needed
for analysis, while all other columns consume space on screen.

So make an option in fselect that allows to choose which columns
to output (like some other databases allows).

Add an option 'columns' in fselect options along with other its
options with list of columns (by number or name). Allow lua tables
and comma-separated-strings for that.

If options argument in fselect is string, interpret is as columns
options.