box: add subscribe infostructure for func cache
There are cases when we need be notified when a function is
added to func cache. At least it is needed for recovery of an
instance with constraints: when need func pointer when loading
space _space, but space _space is loaded earlier than _func and
there are no pointers yet but only function names. We could wait
an set func pointers when space _func is loaded, but the problem
is that constraints are complicated and there's no simple way
to determine by which constraints a func can be used.

Implement func subsrciption for that purpose, and then the func
cache would notify subscriber that a function is loaded and the
func pointer can be used.