ci: split debug_coverage.yml into two workflows
This change splits the debug_coverage.yml workflow into two independent
workflows: debug.yml and coverage.yml.

The first one will be run on push to a branch and creating/updating
external pool requests. In other words, it will be run always.

The second one will be run on push to a branch only if the branch has
the `full-ci` suffix (some-branch-full-ci) or the `full-ci` label is set
on a pool request, internal or external - it doesn't matter.

Why do we need it?

1. The coverage testing is a heavy testing due to enabling long tests
   and takes at least 2 times more time than debug build testing. So
   testing that is done at the first stage (lint, release, debug) will
   pass faster.
2. The coverage report as a pool request comment from coveralls.io can
   be obtained only if a pool request event happened. It is a limitation
   of the coverallsapp/github-action action [1]. So there is no sense to
   run the coverage workflow on push events. But the coverage report on
   push still can be obtained and seen on coveralls.io if the branch has
   the `full-ci` suffix.
3. The logic of the debug_coverage.yml workflow is tricky enough and no
   one understands how it works. So we would like to have an absolutely
   transparent and simple workflow.

[1] https://github.com/coverallsapp/github-action/tree/8cbef1dea373ebce56de0a14c68d6267baa10b44#coveralls-github-action

Closes tarantool/tarantool-qa#147